+++ release/bar-chart/bar-vertical-normalized.component.js
@@ -160,7 +160,7 @@ var BarVerticalNormalizedComponent = (function (_super) {
         });
         this.activeEntries.splice(idx, 1);
         this.activeEntries = this.activeEntries.slice();
-        this.deactivate.emit({ value: event, entries: this.activeEntries });
+        this.deactivate.emit({ value: item, entries: this.activeEntries });
     };
     BarVerticalNormalizedComponent.decorators = [
         { type: core_1.Component, args: [{