+++ release/common/circle-series.component.ngfactory.ts
@@ -15,22 +15,28 @@ import * as import6 from '@angular/core/src/linker/view_type';
 import * as import7 from '@angular/core/src/change_detection/constants';
 import * as import8 from '@angular/core/src/linker/component_factory';
 import * as import9 from '@angular/common/src/location/location';
-import * as import10 from '@angular/core/src/linker/view_container';
-import * as import11 from '../../node_modules/@angular/common/src/directives/ng_for.ngfactory';
-import * as import12 from '@angular/core/src/linker/template_ref';
-import * as import13 from '@angular/core/src/change_detection/differs/iterable_differs';
-import * as import14 from '@angular/common/src/directives/ng_for';
-import * as import15 from '../../../src/common/svg-linear-gradient.component';
-import * as import16 from './svg-linear-gradient.component.ngfactory';
-import * as import17 from '../../node_modules/@angular/common/src/directives/ng_if.ngfactory';
-import * as import18 from '@angular/common/src/directives/ng_if';
-import * as import19 from '../../../src/common/circle.component';
-import * as import20 from './tooltip/tooltip.directive.ngfactory';
-import * as import21 from './circle.component.ngfactory';
-import * as import22 from '../../../src/common/tooltip/tooltip.service';
-import * as import23 from '@angular/core/src/linker/element_ref';
-import * as import24 from '@angular/core/src/zone/ng_zone';
-import * as import25 from '../../../src/common/tooltip/tooltip.directive';
+import * as import10 from '@angular/core/src/animation/animation_transition';
+import * as import11 from '@angular/core/src/animation/animation_sequence_player';
+import * as import12 from '@angular/core/src/animation/animation_styles';
+import * as import13 from '@angular/core/src/animation/animation_style_util';
+import * as import14 from '@angular/core/src/animation/animation_keyframe';
+import * as import15 from '@angular/core/src/animation/animation_player';
+import * as import16 from '@angular/core/src/linker/view_container';
+import * as import17 from '../../node_modules/@angular/common/src/directives/ng_for.ngfactory';
+import * as import18 from '@angular/core/src/linker/template_ref';
+import * as import19 from '@angular/core/src/change_detection/differs/iterable_differs';
+import * as import20 from '@angular/common/src/directives/ng_for';
+import * as import21 from '../../../src/common/svg-linear-gradient.component';
+import * as import22 from './svg-linear-gradient.component.ngfactory';
+import * as import23 from '../../node_modules/@angular/common/src/directives/ng_if.ngfactory';
+import * as import24 from '@angular/common/src/directives/ng_if';
+import * as import25 from '../../../src/common/circle.component';
+import * as import26 from './tooltip/tooltip.directive.ngfactory';
+import * as import27 from './circle.component.ngfactory';
+import * as import28 from '../../../src/common/tooltip/tooltip.service';
+import * as import29 from '@angular/core/src/linker/element_ref';
+import * as import30 from '@angular/core/src/zone/ng_zone';
+import * as import31 from '../../../src/common/tooltip/tooltip.directive';
 export class Wrapper_CircleSeriesComponent {
   /*private*/ _eventHandler:Function;
   context:import0.CircleSeriesComponent;
@@ -187,13 +193,50 @@ class View_CircleSeriesComponent_Host0 extends import2.AppView<any> {
 }
 export const CircleSeriesComponentNgFactory:import8.ComponentFactory<import0.CircleSeriesComponent> = new import8.ComponentFactory<import0.CircleSeriesComponent>('g[ngx-charts-circle-series]',View_CircleSeriesComponent_Host0,import0.CircleSeriesComponent);
 const styles_CircleSeriesComponent:any[] = ([] as any[]);
-var renderType_CircleSeriesComponent:import4.RenderComponentType = import3.createRenderComponentType('',0,import5.ViewEncapsulation.None,styles_CircleSeriesComponent,{});
+var CircleSeriesComponent_animationState_states:any = {
+  '*': {},
+  'void': {}
+}
+;
+function CircleSeriesComponent_animationState_factory(view:import2.AppView<any>,element:any,currentState:any,nextState:any):import10.AnimationTransition {
+  var previousPlayers:any = view.animationContext.getAnimationPlayers(element,((nextState == 'void')? (null as any): 'animationState'));
+  var collectedStyles:any = {};
+  var player:any = (null as any);
+  var totalTime:any = 0;
+  var defaultStateStyles:any = CircleSeriesComponent_animationState_states['*'];
+  var startStateStyles:any = CircleSeriesComponent_animationState_states[currentState];
+  if ((startStateStyles == (null as any))) { (startStateStyles = defaultStateStyles); }
+  var endStateStyles:any = CircleSeriesComponent_animationState_states[nextState];
+  if ((endStateStyles == (null as any))) { (endStateStyles = defaultStateStyles); }
+  if (((player == (null as any)) && ((currentState == 'void') && true))) {
+      player = new import11.AnimationSequencePlayer([view.renderer.animate(element,new import12.AnimationStyles(import13.collectAndResolveStyles(collectedStyles,[
+        startStateStyles,
+        {'opacity': '0'}
+      ]
+      )),[
+        new import14.AnimationKeyframe(0,new import12.AnimationStyles(import13.collectAndResolveStyles(collectedStyles,[{'opacity': '0'}]))),
+        new import14.AnimationKeyframe(1,new import12.AnimationStyles(import13.collectAndResolveStyles(collectedStyles,[{'opacity': '1'}])))
+      ]
+    ,250,0,(null as any),previousPlayers)]);
+    totalTime = 250;
+  }
+  if ((player == (null as any))) { (player = new import15.NoOpAnimationPlayer()); }
+  player.onDone(():void => {
+    player.destroy();
+    import13.renderStyles(element,view.renderer,import13.prepareFinalAnimationStyles(startStateStyles,endStateStyles));
+  });
+  new import11.AnimationSequencePlayer(previousPlayers).destroy();
+  import13.renderStyles(element,view.renderer,import13.clearStyles(startStateStyles));
+  view.animationContext.queueAnimation(element,'animationState',player);
+  return new import10.AnimationTransition(player,currentState,nextState,totalTime);
+}
+var renderType_CircleSeriesComponent:import4.RenderComponentType = import3.createRenderComponentType('',0,import5.ViewEncapsulation.None,styles_CircleSeriesComponent,{animationState: CircleSeriesComponent_animationState_factory});
 export class View_CircleSeriesComponent0 extends import2.AppView<import0.CircleSeriesComponent> {
   _text_0:any;
   _anchor_1:any;
-  /*private*/ _vc_1:import10.ViewContainer;
+  /*private*/ _vc_1:import16.ViewContainer;
   _TemplateRef_1_5:any;
-  _NgFor_1_6:import11.Wrapper_NgFor;
+  _NgFor_1_6:import17.Wrapper_NgFor;
   _text_2:any;
   constructor(viewUtils:import3.ViewUtils,parentView:import2.AppView<any>,parentIndex:number,parentElement:any) {
     super(View_CircleSeriesComponent0,renderType_CircleSeriesComponent,import6.ViewType.COMPONENT,viewUtils,parentView,parentIndex,parentElement,import7.ChangeDetectorStatus.CheckOnce);
@@ -202,9 +245,9 @@ export class View_CircleSeriesComponent0 extends import2.AppView<import0.CircleS
     const parentRenderNode:any = this.renderer.createViewRoot(this.parentElement);
     this._text_0 = this.renderer.createText(parentRenderNode,'\n    ',(null as any));
     this._anchor_1 = this.renderer.createTemplateAnchor(parentRenderNode,(null as any));
-    this._vc_1 = new import10.ViewContainer(1,(null as any),this,this._anchor_1);
-    this._TemplateRef_1_5 = new import12.TemplateRef_(this,1,this._anchor_1);
-    this._NgFor_1_6 = new import11.Wrapper_NgFor(this._vc_1.vcRef,this._TemplateRef_1_5,this.parentView.injectorGet(import13.IterableDiffers,this.parentIndex),this.ref);
+    this._vc_1 = new import16.ViewContainer(1,(null as any),this,this._anchor_1);
+    this._TemplateRef_1_5 = new import18.TemplateRef_(this,1,this._anchor_1);
+    this._NgFor_1_6 = new import17.Wrapper_NgFor(this._vc_1.vcRef,this._TemplateRef_1_5,this.parentView.injectorGet(import19.IterableDiffers,this.parentIndex),this.ref);
     this._text_2 = this.renderer.createText(parentRenderNode,'\n  ',(null as any));
     this.init((null as any),((<any>this.renderer).directRenderer? (null as any): [
       this._text_0,
@@ -215,8 +258,8 @@ export class View_CircleSeriesComponent0 extends import2.AppView<import0.CircleS
     return (null as any);
   }
   injectorGetInternal(token:any,requestNodeIndex:number,notFoundResult:any):any {
-    if (((token === import12.TemplateRef) && (1 === requestNodeIndex))) { return this._TemplateRef_1_5; }
-    if (((token === import14.NgFor) && (1 === requestNodeIndex))) { return this._NgFor_1_6.context; }
+    if (((token === import18.TemplateRef) && (1 === requestNodeIndex))) { return this._TemplateRef_1_5; }
+    if (((token === import20.NgFor) && (1 === requestNodeIndex))) { return this._NgFor_1_6.context; }
     return notFoundResult;
   }
   detectChangesInternal(throwOnChange:boolean):void {
@@ -239,21 +282,21 @@ class View_CircleSeriesComponent1 extends import2.AppView<any> {
   _el_2:any;
   _text_3:any;
   _el_4:any;
-  compView_4:import2.AppView<import15.SvgLinearGradientComponent>;
-  _SvgLinearGradientComponent_4_3:import16.Wrapper_SvgLinearGradientComponent;
+  compView_4:import2.AppView<import21.SvgLinearGradientComponent>;
+  _SvgLinearGradientComponent_4_3:import22.Wrapper_SvgLinearGradientComponent;
   _text_5:any;
   _text_6:any;
   _anchor_7:any;
-  /*private*/ _vc_7:import10.ViewContainer;
+  /*private*/ _vc_7:import16.ViewContainer;
   _TemplateRef_7_5:any;
-  _NgIf_7_6:import17.Wrapper_NgIf;
+  _NgIf_7_6:import23.Wrapper_NgIf;
   _text_8:any;
   _anchor_9:any;
-  /*private*/ _vc_9:import10.ViewContainer;
+  /*private*/ _vc_9:import16.ViewContainer;
   _TemplateRef_9_5:any;
-  _NgIf_9_6:import17.Wrapper_NgIf;
+  _NgIf_9_6:import23.Wrapper_NgIf;
   _text_10:any;
-  constructor(viewUtils:import3.ViewUtils,parentView:import2.AppView<any>,parentIndex:number,parentElement:any,declaredViewContainer:import10.ViewContainer) {
+  constructor(viewUtils:import3.ViewUtils,parentView:import2.AppView<any>,parentIndex:number,parentElement:any,declaredViewContainer:import16.ViewContainer) {
     super(View_CircleSeriesComponent1,renderType_CircleSeriesComponent,import6.ViewType.EMBEDDED,viewUtils,parentView,parentIndex,parentElement,import7.ChangeDetectorStatus.CheckAlways,declaredViewContainer);
   }
   createInternal(rootSelector:string):import8.ComponentRef<any> {
@@ -262,20 +305,20 @@ class View_CircleSeriesComponent1 extends import2.AppView<any> {
     this._el_2 = import3.createRenderElement(this.renderer,this._el_0,':svg:defs',import3.EMPTY_INLINE_ARRAY,(null as any));
     this._text_3 = this.renderer.createText(this._el_2,'\n        ',(null as any));
     this._el_4 = import3.createRenderElement(this.renderer,this._el_2,':svg:g',new import3.InlineArray4(4,'ngx-charts-svg-linear-gradient','','orientation','vertical'),(null as any));
-    this.compView_4 = new import16.View_SvgLinearGradientComponent0(this.viewUtils,this,4,this._el_4);
-    this._SvgLinearGradientComponent_4_3 = new import16.Wrapper_SvgLinearGradientComponent();
+    this.compView_4 = new import22.View_SvgLinearGradientComponent0(this.viewUtils,this,4,this._el_4);
+    this._SvgLinearGradientComponent_4_3 = new import22.Wrapper_SvgLinearGradientComponent();
     this.compView_4.create(this._SvgLinearGradientComponent_4_3.context);
     this._text_5 = this.renderer.createText(this._el_2,'\n      ',(null as any));
     this._text_6 = this.renderer.createText(this._el_0,'\n      ',(null as any));
     this._anchor_7 = this.renderer.createTemplateAnchor(this._el_0,(null as any));
-    this._vc_7 = new import10.ViewContainer(7,0,this,this._anchor_7);
-    this._TemplateRef_7_5 = new import12.TemplateRef_(this,7,this._anchor_7);
-    this._NgIf_7_6 = new import17.Wrapper_NgIf(this._vc_7.vcRef,this._TemplateRef_7_5);
+    this._vc_7 = new import16.ViewContainer(7,0,this,this._anchor_7);
+    this._TemplateRef_7_5 = new import18.TemplateRef_(this,7,this._anchor_7);
+    this._NgIf_7_6 = new import23.Wrapper_NgIf(this._vc_7.vcRef,this._TemplateRef_7_5);
     this._text_8 = this.renderer.createText(this._el_0,'\n      ',(null as any));
     this._anchor_9 = this.renderer.createTemplateAnchor(this._el_0,(null as any));
-    this._vc_9 = new import10.ViewContainer(9,0,this,this._anchor_9);
-    this._TemplateRef_9_5 = new import12.TemplateRef_(this,9,this._anchor_9);
-    this._NgIf_9_6 = new import17.Wrapper_NgIf(this._vc_9.vcRef,this._TemplateRef_9_5);
+    this._vc_9 = new import16.ViewContainer(9,0,this,this._anchor_9);
+    this._TemplateRef_9_5 = new import18.TemplateRef_(this,9,this._anchor_9);
+    this._NgIf_9_6 = new import23.Wrapper_NgIf(this._vc_9.vcRef,this._TemplateRef_9_5);
     this._text_10 = this.renderer.createText(this._el_0,'\n    ',(null as any));
     this.init(this._el_0,((<any>this.renderer).directRenderer? (null as any): [
       this._el_0,
@@ -294,11 +337,11 @@ class View_CircleSeriesComponent1 extends import2.AppView<any> {
     return (null as any);
   }
   injectorGetInternal(token:any,requestNodeIndex:number,notFoundResult:any):any {
-    if (((token === import15.SvgLinearGradientComponent) && (4 === requestNodeIndex))) { return this._SvgLinearGradientComponent_4_3.context; }
-    if (((token === import12.TemplateRef) && (7 === requestNodeIndex))) { return this._TemplateRef_7_5; }
-    if (((token === import18.NgIf) && (7 === requestNodeIndex))) { return this._NgIf_7_6.context; }
-    if (((token === import12.TemplateRef) && (9 === requestNodeIndex))) { return this._TemplateRef_9_5; }
-    if (((token === import18.NgIf) && (9 === requestNodeIndex))) { return this._NgIf_9_6.context; }
+    if (((token === import21.SvgLinearGradientComponent) && (4 === requestNodeIndex))) { return this._SvgLinearGradientComponent_4_3.context; }
+    if (((token === import18.TemplateRef) && (7 === requestNodeIndex))) { return this._TemplateRef_7_5; }
+    if (((token === import24.NgIf) && (7 === requestNodeIndex))) { return this._NgIf_7_6.context; }
+    if (((token === import18.TemplateRef) && (9 === requestNodeIndex))) { return this._TemplateRef_9_5; }
+    if (((token === import24.NgIf) && (9 === requestNodeIndex))) { return this._NgIf_9_6.context; }
     return notFoundResult;
   }
   detectChangesInternal(throwOnChange:boolean):void {
@@ -342,13 +385,15 @@ class View_CircleSeriesComponent2 extends import2.AppView<any> {
   /*private*/ _expr_3:any;
   /*private*/ _expr_4:any;
   /*private*/ _expr_5:any;
-  constructor(viewUtils:import3.ViewUtils,parentView:import2.AppView<any>,parentIndex:number,parentElement:any,declaredViewContainer:import10.ViewContainer) {
+  /*private*/ _expr_6:any;
+  constructor(viewUtils:import3.ViewUtils,parentView:import2.AppView<any>,parentIndex:number,parentElement:any,declaredViewContainer:import16.ViewContainer) {
     super(View_CircleSeriesComponent2,renderType_CircleSeriesComponent,import6.ViewType.EMBEDDED,viewUtils,parentView,parentIndex,parentElement,import7.ChangeDetectorStatus.CheckAlways,declaredViewContainer);
     this._expr_1 = import1.UNINITIALIZED;
     this._expr_2 = import1.UNINITIALIZED;
     this._expr_3 = import1.UNINITIALIZED;
     this._expr_4 = import1.UNINITIALIZED;
     this._expr_5 = import1.UNINITIALIZED;
+    this._expr_6 = import1.UNINITIALIZED;
   }
   createInternal(rootSelector:string):import8.ComponentRef<any> {
     this._el_0 = import3.createRenderElement(this.renderer,(null as any),':svg:rect',new import3.InlineArray2(2,'class','tooltip-bar'),(null as any));
@@ -356,31 +401,39 @@ class View_CircleSeriesComponent2 extends import2.AppView<any> {
     return (null as any);
   }
   detectChangesInternal(throwOnChange:boolean):void {
-    const currVal_1:any = (this.parentView.context.$implicit.cx - this.parentView.context.$implicit.radius);
+    const currVal_1:any = 'active';
     if (import3.checkBinding(throwOnChange,this._expr_1,currVal_1)) {
-      this.renderer.setElementAttribute(this._el_0,'x',((currVal_1 == null)? (null as any): currVal_1.toString()));
+      var animationTransition_animationState:any = this.componentType.animations['animationState'](this,this._el_0,((this._expr_1 == import1.UNINITIALIZED)? 'void': this._expr_1),((currVal_1 == import1.UNINITIALIZED)? 'void': currVal_1));
       this._expr_1 = currVal_1;
     }
-    const currVal_2:any = this.parentView.context.$implicit.cy;
+    const currVal_2:any = (this.parentView.context.$implicit.cx - this.parentView.context.$implicit.radius);
     if (import3.checkBinding(throwOnChange,this._expr_2,currVal_2)) {
-      this.renderer.setElementAttribute(this._el_0,'y',((currVal_2 == null)? (null as any): currVal_2.toString()));
+      this.renderer.setElementAttribute(this._el_0,'x',((currVal_2 == null)? (null as any): currVal_2.toString()));
       this._expr_2 = currVal_2;
     }
-    const currVal_3:any = (this.parentView.context.$implicit.radius * 2);
+    const currVal_3:any = this.parentView.context.$implicit.cy;
     if (import3.checkBinding(throwOnChange,this._expr_3,currVal_3)) {
-      this.renderer.setElementAttribute(this._el_0,'width',((currVal_3 == null)? (null as any): currVal_3.toString()));
+      this.renderer.setElementAttribute(this._el_0,'y',((currVal_3 == null)? (null as any): currVal_3.toString()));
       this._expr_3 = currVal_3;
     }
-    const currVal_4:any = this.parentView.context.$implicit.height;
+    const currVal_4:any = (this.parentView.context.$implicit.radius * 2);
     if (import3.checkBinding(throwOnChange,this._expr_4,currVal_4)) {
-      this.renderer.setElementAttribute(this._el_0,'height',((currVal_4 == null)? (null as any): currVal_4.toString()));
+      this.renderer.setElementAttribute(this._el_0,'width',((currVal_4 == null)? (null as any): currVal_4.toString()));
       this._expr_4 = currVal_4;
     }
-    const currVal_5:any = this.parentView.context.$implicit.gradientFill;
+    const currVal_5:any = this.parentView.context.$implicit.height;
     if (import3.checkBinding(throwOnChange,this._expr_5,currVal_5)) {
-      this.renderer.setElementAttribute(this._el_0,'fill',((currVal_5 == null)? (null as any): currVal_5.toString()));
+      this.renderer.setElementAttribute(this._el_0,'height',((currVal_5 == null)? (null as any): currVal_5.toString()));
       this._expr_5 = currVal_5;
     }
+    const currVal_6:any = this.parentView.context.$implicit.gradientFill;
+    if (import3.checkBinding(throwOnChange,this._expr_6,currVal_6)) {
+      this.renderer.setElementAttribute(this._el_0,'fill',((currVal_6 == null)? (null as any): currVal_6.toString()));
+      this._expr_6 = currVal_6;
+    }
+  }
+  detachInternal():void {
+    var animationTransition_animationState:any = this.componentType.animations['animationState'](this,this._el_0,this._expr_1,'void');
   }
   visitRootNodesInternal(cb:any,ctx:any):void {
     cb(this._el_0,ctx);
@@ -388,14 +441,14 @@ class View_CircleSeriesComponent2 extends import2.AppView<any> {
 }
 class View_CircleSeriesComponent3 extends import2.AppView<any> {
   _el_0:any;
-  /*private*/ _vc_0:import10.ViewContainer;
-  compView_0:import2.AppView<import19.CircleComponent>;
-  _TooltipDirective_0_5:import20.Wrapper_TooltipDirective;
-  _CircleComponent_0_6:import21.Wrapper_CircleComponent;
+  /*private*/ _vc_0:import16.ViewContainer;
+  compView_0:import2.AppView<import25.CircleComponent>;
+  _TooltipDirective_0_5:import26.Wrapper_TooltipDirective;
+  _CircleComponent_0_6:import27.Wrapper_CircleComponent;
   _el_1:any;
   /*private*/ _expr_6:any;
   _map_7:any;
-  constructor(viewUtils:import3.ViewUtils,parentView:import2.AppView<any>,parentIndex:number,parentElement:any,declaredViewContainer:import10.ViewContainer) {
+  constructor(viewUtils:import3.ViewUtils,parentView:import2.AppView<any>,parentIndex:number,parentElement:any,declaredViewContainer:import16.ViewContainer) {
     super(View_CircleSeriesComponent3,renderType_CircleSeriesComponent,import6.ViewType.EMBEDDED,viewUtils,parentView,parentIndex,parentElement,import7.ChangeDetectorStatus.CheckAlways,declaredViewContainer);
     this._expr_6 = import1.UNINITIALIZED;
     this._map_7 = import3.pureProxy1((p0:any):{[key: string]:any} => {
@@ -404,10 +457,10 @@ class View_CircleSeriesComponent3 extends import2.AppView<any> {
   }
   createInternal(rootSelector:string):import8.ComponentRef<any> {
     this._el_0 = import3.createRenderElement(this.renderer,(null as any),':svg:g',new import3.InlineArray8(6,'class','circle','ngx-charts-circle','','ngx-tooltip',''),(null as any));
-    this._vc_0 = new import10.ViewContainer(0,(null as any),this,this._el_0);
-    this.compView_0 = new import21.View_CircleComponent0(this.viewUtils,this,0,this._el_0);
-    this._TooltipDirective_0_5 = new import20.Wrapper_TooltipDirective(this.parentView.parentView.injectorGet(import22.TooltipService,this.parentView.parentIndex),this._vc_0.vcRef,this.renderer,new import23.ElementRef(this._el_0),this.parentView.parentView.injectorGet(import24.NgZone,this.parentView.parentIndex));
-    this._CircleComponent_0_6 = new import21.Wrapper_CircleComponent();
+    this._vc_0 = new import16.ViewContainer(0,(null as any),this,this._el_0);
+    this.compView_0 = new import27.View_CircleComponent0(this.viewUtils,this,0,this._el_0);
+    this._TooltipDirective_0_5 = new import26.Wrapper_TooltipDirective(this.parentView.parentView.injectorGet(import28.TooltipService,this.parentView.parentIndex),this._vc_0.vcRef,this.renderer,new import29.ElementRef(this._el_0),this.parentView.parentView.injectorGet(import30.NgZone,this.parentView.parentIndex));
+    this._CircleComponent_0_6 = new import27.Wrapper_CircleComponent();
     this.compView_0.create(this._CircleComponent_0_6.context);
     this._el_1 = this.renderer.createTemplateAnchor((null as any),(null as any));
     var disposable_0:Function = import3.subscribeToRenderElement(this,this._el_0,new import3.InlineArray16(16,'select',(null as any),'activate',(null as any),'deactivate',(null as any),'focusin',(null as any),'blur',(null as any),'mouseenter',(null as any),'mouseleave',(null as any),'click',(null as any)),this.eventHandler(this.handleEvent_0));
@@ -416,8 +469,8 @@ class View_CircleSeriesComponent3 extends import2.AppView<any> {
     return (null as any);
   }
   injectorGetInternal(token:any,requestNodeIndex:number,notFoundResult:any):any {
-    if (((token === import25.TooltipDirective) && (0 === requestNodeIndex))) { return this._TooltipDirective_0_5.context; }
-    if (((token === import19.CircleComponent) && (0 === requestNodeIndex))) { return this._CircleComponent_0_6.context; }
+    if (((token === import31.TooltipDirective) && (0 === requestNodeIndex))) { return this._TooltipDirective_0_5.context; }
+    if (((token === import25.CircleComponent) && (0 === requestNodeIndex))) { return this._CircleComponent_0_6.context; }
     return notFoundResult;
   }
   detectChangesInternal(throwOnChange:boolean):void {